style(7): fix all clang-tidy warnings in production sources
Suppress 3 false-positive categories in .clang-tidy:
- bugprone-too-small-loop-variable (seqan API size_t vs uint16_t)
- bugprone-inc-dec-in-conditions (intentional while(n--) pattern)
- performance-enum-size (not worth changing enum layout)

Fix 87 warnings across 30 files:
- Remove 44 unused using declarations
- Add braces around single-line statements
- Convert push_back to emplace_back
- Convert index-based loops to range-for
- Add static_cast<size_t>() for 19 widening multiplications
- Add default:break to 8 switch statements
- Fix 3 pass-by-value to const-ref (stop_token, Dna5String)
- Fix 2 integer divisions in floating-point contexts
- Initialize uninitialized bool variable and 2 struct fields
- Rename reserved identifier (double underscore) in format_utils
- Add override keyword to virtual function
- Add NOLINT for 3 false-positive static analyzer warnings

Remove continue-on-error from lint CI job -- now passes cleanly.

+++ reseq/Surrounding.h
@@ -110,8 +110,8 @@ class SurroundingBias {
         }
     }
 
-    void CombinePositions(const std::array<double, 4 * Surrounding::Length()>& separated);
-    void SeparatePositions(std::array<double, 4 * Surrounding::Length()>& separated) const;
+    void CombinePositions(const std::array<double, static_cast<size_t>(4) * Surrounding::Length()>& separated);
+    void SeparatePositions(std::array<double, static_cast<size_t>(4) * Surrounding::Length()>& separated) const;
 
     double Bias(const Surrounding& sur) const {
         double bias = 0.0;
